e-reputation/requestanalysis/app/requirements.txt
flask
textblob
kafka-python
requests